Fix active field for factory retailer relation

+++ src/Furniture/FrontendBundle/Controller/Profile/Retailer/FactoryRelationController.php
@@ -266,6 +266,13 @@ class FactoryRelationController
             throw new AccessDeniedException();
         }
 
+        if ($relation->isNotActive()) {
+            throw new NotFoundHttpException(sprintf(
+                'The relation with id "%s" is not active.',
+                $relation->getId()
+            ));
+        }
+
         $this->em->remove($relation);
         $this->em->flush();
